NEW11+NEW12+NEW13: półgwiazdki, filtr ocen, przycisk Dom na mapie
- rating: INTEGER → NUMERIC(2,1), krok 0.5 (0.5–5.0); migracja w ensure_schema()
- stars() renderuje pół-gwiazdę przez background-clip:text
- nowy sort-bar w liście podróży: Wszystkie / 5 / ≥4.5 / ≥4 / ≥3 / Bez oceny
- przycisk 🏠 obok 🔄 na mapie centruje na Osielsku (HOME_COORDS)

+++ migrate.py
@@ -72,7 +72,7 @@ CREATE TABLE travels (
     amount NUMERIC(12,2),
     currency TEXT DEFAULT 'PLN',
     is_description_complete BOOLEAN DEFAULT FALSE,
-    rating INTEGER,
+    rating NUMERIC(2,1),
     reflections TEXT,
     notes TEXT,
     number_of_flights INTEGER DEFAULT 0
